backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
openai>=1.0.0
pypdf2>=3.0.0
python-docx>=0.8.11
python-dotenv>=0.19.0